requirements.txt
opencv-python>=4.1.0.25
numpy>=1.16.4
pdf2image>=1.5.4
Pillow>=6.0.0
PyQt5>=5.12.1
fpdf2>=2.0.3